[Examples] Reorganize Blackwell matmul examples V2-V6
Reorganize the Blackwell matmul tutorial examples so each version
introduces exactly one major optimization over the previous:

| Version | New Optimization                              | Warps | Cluster |
|---------|-----------------------------------------------|-------|---------|
| V0      | Baseline (copy_async + tcgen05 MMA)           | 4     | No      |
| V1      | TMA loads                                     | 4     | No      |
| V2      | Software pipelining (multi-stage)              | 4     | No      |
| V3      | Warp specialization (TMA warp + MMA warp)     | 4     | No      |
| V4      | Pipeline class + TMA epilogue + L2 swizzle    | 4     | No      |
| V5      | CLC persistent kernel + pipelined epilogue     | 8     | 1 CTA   |
| V6      | 2-CTA cluster + distributed MMA (cta_group=2) | 8     | 2 CTAs  |

Changes:
- Rewrite V4: inline Pipeline abstraction, TMA epilogue store (replaces
  direct store_global), L2 tile swizzle (compute_block_coord)
- Rewrite V5: CLC persistent kernel with single-CTA cluster, pipelined
  epilogue via mma_pipe, 4 specialized warp roles (TMA/MMA/scheduler/
  epilogue), two-level pipeline (tma_pipe + mma_pipe)
- Rewrite V6: add 2-CTA distributed MMA on top of V5 (cta_group=2,
  multicast barriers, cross-CTA shared memory access)
- Remove V7 and V8 (consolidated into V5 and V6)
- Add relaxed memory ordering (sem="relaxed") to mbarrier.wait in V2-V6
- Add tma.wait_group(read=True) hint in V4-V6
- Fix codegen to always emit __cluster_dims__ even for (1,1,1), enabling
  CLC with single-CTA clusters
